requirements.txt
flask_cors
Flask==3.0.2
numpy==1.23.4
networkx==2.8.8
PyYAML==6.0
redis==4.6.0
matplotlib~=3.8.3
simpy
python-louvain==0.16
scipy
psycopg2-binary
requests
